requirements.txt
Django==4.2.9
Pillow==10.2.0
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-taggit==4.0.0
whitenoise==6.5.0
gunicorn==21.2.0
